feat: jwt 필터에서 userId 타입을 long에서 int로 변경

+++ src/main/java/com/example/outsourcingproject/global/filter/JwtFilter.java
@@ -55,7 +55,7 @@ public class JwtFilter implements Filter {
                 return;
             }
 
-            httpRequest.setAttribute("id", Long.parseLong(claims.getSubject()));
+            httpRequest.setAttribute("id", Integer.parseInt(claims.getSubject()));
             httpRequest.setAttribute("email", claims.get("email"));
